Reorganize code
Separate the "dialogue runtime" which tracks conversation histories/APIs/logs
and is independent of ThingTalk/dialogue PL, from the "thingtalk dialogues"
(dialogue PL)
