+++ lib/dialogue-runtime/card-output/card-formatter.ts
