+++ lib/dialogue-runtime/errors.ts
