+++ lib/dialogue-runtime/notifications/formatter.ts
@@ -25,8 +25,8 @@ import * as ThingTalk from 'thingtalk';
 import Engine from '../../engine';
 import * as I18n from '../../i18n';
 
-import StatementExecutor from '../statement_executor';
-import { DialoguePolicy } from '../../new-dialogue-agent/inference-time-dialogue';
+import StatementExecutor from '../../thingtalk-dialogues/statement_executor';
+import { DialoguePolicy } from '../../thingtalk-dialogues/inference-time-dialogue';
 
 import CardFormatter from '../card-output/card-formatter';
 import { FORMAT_TYPES } from '../card-output/format_objects';
