+++ lib/dialogue-runtime/protocol.ts
