+++ lib/dialogue-runtime/value-category.ts
