+++ lib/thingtalk-dialogues/cmd-dispatch.ts
