+++ lib/thingtalk-dialogues/command.ts
