+++ lib/thingtalk-dialogues/simulator/helpers.ts
