+++ lib/thingtalk-dialogues/simulator/simulation_exec_environment.ts
