+++ lib/thingtalk-dialogues/simulator/statement_simulator.ts
