+++ lib/thingtalk-dialogues/simulator/types.ts
