--- test/unit/test_helpers.js
@@ -1,62 +1,0 @@
-// -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
-//
-// This file is part of Genie
-//
-// Copyright 2020 The Board of Trustees of the Leland Stanford Junior University
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
-
-
-import * as Tp from 'thingpedia';
-import * as ThingTalk from 'thingtalk';
-
-import { MockPlatform } from './mock_utils';
-import * as Helpers from '../../lib/dialogue-agent/helpers';
-
-const CATEGORIES = ['media', 'social-network', 'home', 'communication', 'health', 'service', 'data-management'];
-
-
-async function test(tpClient) {
-    const schemas = new ThingTalk.SchemaRetriever(tpClient, null, true);
-
-    for (const category of CATEGORIES) {
-        console.log('category', category);
-
-        for (let page = 0; ; page++) {
-            const devices = await tpClient.getDeviceList(category, page, 10);
-            let hasMore = false;
-            if (devices.length > 10) {
-                hasMore = true;
-                devices.length = 10;
-            }
-            for (let device of devices) {
-                console.log('device', '@' + device.primary_kind);
-                const examples = await tpClient.getExamplesByKinds([device.primary_kind], true);
-                await Helpers.loadExamples(examples, schemas);
-            }
-            if (!hasMore)
-                break;
-        }
-    }
-}
-
-async function main() {
-    const platform = new MockPlatform();
-    const tpClient = new Tp.HttpClient(platform, 'https://dev.almond.stanford.edu/thingpedia');
-    await test(tpClient);
-}
-export default main;
-if (!module.parent)
-    main();
